runner: it's alive!

+++ runner/teardown.go
@@ -11,9 +11,11 @@ func (m *Match) teardown() (err error) {
 
 	for name, process := range m.Players {
 		m.logger.Debug("Killing player", "player", name)
-		err := process.Kill()
-		if err != nil {
-			m.logger.Error("Could not kill player", "err", err)
+		if process.IsRunning() && process.Process != nil {
+			err := process.Kill()
+			if err != nil {
+				m.logger.Error("Could not kill player", "err", err)
+			}
 		}
 	}
